Re-organize NIO classes
Fixes #11
