+++ src/main/java/com/rabbitmq/client/impl/nio/SocketChannelFrameHandlerState.java
@@ -1,5 +1,22 @@
-package com.rabbitmq.client.impl;
+// Copyright (c) 2007-Present Pivotal Software, Inc.  All rights reserved.
+//
+// This software, the RabbitMQ Java client library, is triple-licensed under the
+// Mozilla Public License 1.1 ("MPL"), the GNU General Public License version 2
+// ("GPL") and the Apache License version 2 ("ASL"). For the MPL, please see
+// LICENSE-MPL-RabbitMQ. For the GPL, please see LICENSE-GPL2.  For the ASL,
+// please see LICENSE-APACHE2.
+//
+// This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY KIND,
+// either express or implied. See the LICENSE file for specific language governing
+// rights and limitations of this software.
+//
+// If you have any questions regarding licensing, please contact us at
+// info@rabbitmq.com.
 
+package com.rabbitmq.client.impl.nio;
+
+import com.rabbitmq.client.impl.AMQConnection;
+import com.rabbitmq.client.impl.Frame;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -9,7 +26,6 @@ import java.nio.channels.SocketChannel;
 import java.util.Queue;
 import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.TimeUnit;
 
 /**
@@ -30,14 +46,14 @@ public class SocketChannelFrameHandlerState {
     /** should be used only in the NIO read thread */
     private long lastActivity;
 
-    private final SocketChannelFrameHandlerFactory.SelectorState writeSelectorState;
+    private final SelectorHolder writeSelectorState;
 
-    private final SocketChannelFrameHandlerFactory.SelectorState readSelectorState;
+    private final SelectorHolder readSelectorState;
 
     private final int writeEnqueuingTimeoutInMs;
 
-    public SocketChannelFrameHandlerState(SocketChannel channel, SocketChannelFrameHandlerFactory.SelectorState readSelectorState,
-        SocketChannelFrameHandlerFactory.SelectorState writeSelectorState, NioParams nioParams) {
+    public SocketChannelFrameHandlerState(SocketChannel channel, SelectorHolder readSelectorState,
+        SelectorHolder writeSelectorState, NioParams nioParams) {
         this.channel = channel;
         this.readSelectorState = readSelectorState;
         this.writeSelectorState = writeSelectorState;
