+++ src/test/java/com/rabbitmq/client/test/functional/ConnectionRecovery.java
@@ -16,7 +16,7 @@
 package com.rabbitmq.client.test.functional;
 
 import com.rabbitmq.client.*;
-import com.rabbitmq.client.impl.NioParams;
+import com.rabbitmq.client.impl.nio.NioParams;
 import com.rabbitmq.client.impl.recovery.*;
 import com.rabbitmq.client.test.BrokerTestCase;
 import com.rabbitmq.tools.Host;
